requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
Pillow>=10.0
torch>=2.3
diffusers @ git+https://github.com/huggingface/diffusers
transformers>=4.45
accelerate>=0.30
huggingface_hub>=0.24
